Python_Agent_Hub/requirements.txt
# === CORE FRAMEWORK ===
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# === DATA PROCESSING ===
pandas>=2.0.0
numpy>=1.24.0

# === FUNDAMENTALS & MACRO DATA ===
fredapi>=0.5.2
yfinance>=0.2.40

# === NEWS & SENTIMENT (FinBERT) ===
transformers>=4.35.0
torch>=2.1.0
scipy>=1.11.0

# === ASYNC SUPPORT ===
aiohttp>=3.9.0
httpx>=0.25.0

# === WEB SCRAPING (Forex Factory) ===
beautifulsoup4>=4.12.0
